+++ components/TxForm.tsx
@@ -4,7 +4,27 @@ import { Input, InputGroup, InputRightElement } from "@chakra-ui/input";
 import { Box, Flex } from "@chakra-ui/layout";
 import { Select } from "@chakra-ui/select";
 
-export const TxForm = ({ tokenList }: { tokenList: { name: string }[] }) => {
+// TODO this is an example list, update token list with real list of supported tokens
+const tokenList = [
+  {
+    chainId: 1,
+    address: "0x111111111117dC0aa78b770fA6A738034120C302",
+    name: "1inch",
+    symbol: "1INCH",
+    decimals: 18,
+    logoURI: "https://assets.coingecko…nch-token.png?1608803028",
+  },
+  {
+    chainId: 1,
+    address: "0x7Fc66500c84A76Ad7e9c93437bFc5Ac33E2DDaE9",
+    name: "Aave",
+    symbol: "AAVE",
+    decimals: 18,
+    logoURI: "https://assets.coingecko…humb/AAVE.png?1601374110",
+  },
+];
+
+export const TxForm = () => {
   return (
     <Box width="24rem">
       <FormControl>
